+++ src/lib/index.js
@@ -1,20 +1,7 @@
-// Utils
 export * from './utils.js';
-
-// Constants
 export * from './constants.js';
-
-// Whitespace
-export * from './whitespace.js';
-
-// Attributes
-export * from './attributes.js';
-
-// Elements
 export * from './elements.js';
-
-// Content processors
+export * from './attributes.js';
 export * from './content.js';
-
-// Options
+export * from './whitespace.js';
 export * from './options.js';