+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [4.16.5] - 2026-01-@@
 
+### Changed
+
+- More JavaScript-related `type` attributes are being removed (`executableScriptsMimetypes`)
+
 ### Internal
 
 - Reformatted and optimized code and documentation